# docs/migration/phase1/postgres-openapi-baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Learnova Backend",
    "description": "Backend API for Learnova learner and instructor workflows.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Courses",
        "description": "This returns the learner dashboard payload for the authenticated user.",
        "operationId": "list_courses_courses_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/courses/{course_slug}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Course",
        "description": "This returns the learner-facing detail payload for one course.",
        "operationId": "get_course_courses__course_slug__get",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/courses/{course_slug}/content/{content_slug}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Course Content",
        "description": "This returns a single content item payload for direct lesson-player loading.",
        "operationId": "get_course_content_courses__course_slug__content__content_slug__get",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "content_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/courses/{course_slug}/quizzes/{content_slug}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Quiz",
        "description": "This returns the quiz payload for the requested course content slug.",
        "operationId": "get_quiz_courses__course_slug__quizzes__content_slug__get",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "content_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/reports/course-progress": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Course Progress Report",
        "description": "This returns reporting summary cards and course-wise learner progress rows.",
        "operationId": "get_course_progress_report_admin_reports_course_progress_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Admin Users",
        "description": "This returns admin/instructor user records for responsible-user dropdowns.",
        "operationId": "get_admin_users_admin_users_get",
        "parameters": [
          {
            "name": "roles",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Roles"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/uploads": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Upload Admin File",
        "description": "This stores a local uploaded asset and returns its served URL.",
        "operationId": "upload_admin_file_admin_uploads_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_admin_file_admin_uploads_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/courses": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Admin Courses",
        "description": "This returns the instructor/admin course list for backoffice views.",
        "operationId": "get_admin_courses_admin_courses_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Post Admin Course",
        "description": "This creates a new course record and attaches its initial tags.",
        "operationId": "post_admin_course_admin_courses_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/courses/{course_slug}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Admin Course Detail",
        "description": "This returns one course record for the instructor course form.",
        "operationId": "get_admin_course_detail_admin_courses__course_slug__get",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Put Admin Course",
        "description": "This updates the editable course fields and course tags.",
        "operationId": "put_admin_course_admin_courses__course_slug__put",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Course",
        "description": "This deletes a course and its dependent rows through database cascades.",
        "operationId": "delete_course_admin_courses__course_slug__delete",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/courses/{course_slug}/publish": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Publish Course",
        "description": "This toggles the publish state used by the website/app visibility flow.",
        "operationId": "publish_course_admin_courses__course_slug__publish_post",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishCourseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/courses/{course_slug}/attendees": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Attendees",
        "description": "This returns the attendee list for the selected course.",
        "operationId": "get_attendees_admin_courses__course_slug__attendees_get",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Post Attendees",
        "description": "This adds or updates attendee enrollments and supports invitation/admin-added flows.",
        "operationId": "post_attendees_admin_courses__course_slug__attendees_post",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAttendeesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddAttendeesRequest": {
        "properties": {
          "attendees": {
            "items": {
              "$ref": "#/components/schemas/AttendeeInviteItem"
            },
            "type": "array",
            "title": "Attendees"
          }
        },
        "type": "object",
        "required": [
          "attendees"
        ],
        "title": "AddAttendeesRequest"
      },
...
      "Body_upload_admin_file_admin_uploads_post": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category",
            "default": "attachments"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_admin_file_admin_uploads_post"
      },
...
      "CourseCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Title"
          },
          "shortDescription": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Shortdescription"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "thumbnailUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailurl"
          },
          "coverImageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverimageurl"
          },
          "websiteId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Websiteid"
          },
          "visibility": {
            "type": "string",
            "pattern": "^(everyone|signed_in)$",
            "title": "Visibility",
            "default": "everyone"
          },
          "accessRule": {
            "type": "string",
            "pattern": "^(open|invitation|payment)$",
            "title": "Accessrule",
            "default": "open"
          },
          "price": {
            "type": "number",
            "minimum": 0.0,
            "title": "Price",
            "default": 0
          },
          "responsibleUserId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsibleuserid"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "isPublished": {
            "type": "boolean",
            "title": "Ispublished",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "title",
          "shortDescription"
        ],
        "title": "CourseCreateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublishCourseRequest": {
        "properties": {
          "isPublished": {
            "type": "boolean",
            "title": "Ispublished"
          }
        },
        "type": "object",
        "required": [
          "isPublished"
        ],
        "title": "PublishCourseRequest"
      },
...
    }
  }
}